@@ -1,7 +1,6 @@
 package godot.gradle
 
 import godot.codegen.generateApiFrom
-import godot.docgen.DocGen
 import org.gradle.api.DefaultTask
 import org.gradle.api.Plugin
 import org.gradle.api.Project
@@ -10,8 +9,6 @@ import org.gradle.api.tasks.InputDirectory
 import org.gradle.api.tasks.InputFile
 import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.TaskAction
-import org.gradle.kotlin.dsl.create
-import org.gradle.kotlin.dsl.property
 
 open class ApiGeneratorPluginExtension(objects: ObjectFactory) {
     var outputDir = objects.directoryProperty()
@@ -38,14 +35,14 @@ open class GenerateAPI : DefaultTask() {
 
 class ApiGeneratorPlugin : Plugin<Project> {
     override fun apply(project: Project) {
-        val extension = project.extensions.create<ApiGeneratorPluginExtension>("apiGenerator")
-        project.tasks.register("generateAPI", GenerateAPI::class.java) {
-            outputDir.set(extension.outputDir)
-            sourceJson.set(extension.sourceJson)
-            docsDir.set(extension.docsDir)
+        val extension = project.extensions.create("apiGenerator", ApiGeneratorPluginExtension::class.java)
+        project.tasks.register("generateAPI", GenerateAPI::class.java) { task ->
+            task.outputDir.set(extension.outputDir)
+            task.sourceJson.set(extension.sourceJson)
+            task.docsDir.set(extension.docsDir)
 
-            group = "godot-jvm"
-            description = "Generate Godot's classes from its api."
+            task.group = "godot-jvm"
+            task.description = "Generate Godot's classes from its api."
         }
     }
 }
