@@ -18,12 +18,12 @@ fun Project.setupConfigurationsAndCompilations() {
     //add our dependencies to the main compilation -> convenience for the user
     kotlinJvmExtension.target.compilations.getByName("main").apply {
         dependencies {
-            compileOnly("com.utopia-rise:godot-library:${GodotBuildProperties.godotKotlinVersion}")
-            compileOnly("com.utopia-rise:godot-kotlin-symbol-processor:${GodotBuildProperties.godotKotlinVersion}")
+            compileOnly("com.utopia-rise:godot-library:${GodotBuildProperties.godotKotlinJvmVersion}")
+            compileOnly("com.utopia-rise:godot-kotlin-symbol-processor:${GodotBuildProperties.godotKotlinJvmVersion}")
         }
         dependencies.add(
             "ksp",
-            "com.utopia-rise:godot-kotlin-symbol-processor:${GodotBuildProperties.godotKotlinVersion}"
+            "com.utopia-rise:godot-kotlin-symbol-processor:${GodotBuildProperties.godotKotlinJvmVersion}"
         )
     }
 
@@ -32,7 +32,7 @@ fun Project.setupConfigurationsAndCompilations() {
     val bootstrapConfiguration = configurations.create("bootstrap") {
         with(it.dependencies) {
             add(dependencies.create("org.jetbrains.kotlin:kotlin-stdlib:${kotlinJvmExtension.coreLibrariesVersion}"))
-            add(dependencies.create("com.utopia-rise:godot-library:${GodotBuildProperties.godotKotlinVersion}"))
+            add(dependencies.create("com.utopia-rise:godot-library:${GodotBuildProperties.godotKotlinJvmVersion}"))
         }
     }
 }
