@@ -1,6 +1,5 @@
 package godot.intellij.plugin.annotator.property
 
-import godot.intellij.plugin.extension.getType
 import godot.tools.common.constants.KotlinFunctions
 import godot.tools.common.constants.kotlinCollectionsPackage
 import godot.tools.common.constants.kotlinTextPackage
@@ -21,6 +20,7 @@ import org.jetbrains.kotlin.psi.KtOperationReferenceExpression
 import org.jetbrains.kotlin.psi.KtStringTemplateExpression
 import org.jetbrains.kotlin.psi.psiUtil.referenceExpression
 import org.jetbrains.kotlin.resolve.bindingContextUtil.getReferenceTargets
+import org.jetbrains.kotlin.resolve.calls.util.getType
 import org.jetbrains.kotlin.resolve.descriptorUtil.fqNameSafe
 import org.jetbrains.kotlin.resolve.descriptorUtil.isCompanionObject
 import org.jetbrains.kotlin.resolve.lazy.BodyResolveMode
